moq-relay: move /metrics to its own `metrics` module + listener
Replace the `--web-internal-listen` field on WebConfig with a standalone
`metrics` module: a `MetricsConfig` (`--metrics-listen` / MOQ_METRICS_LISTEN)
peer to `web`/`stats` in the top-level Config, plus a `Metrics` service that
serves `/metrics` (+ a `/health` mirror) on its own plain-HTTP listener over
the node's Stats handle.

This is a cleaner seam than folding it into the customer-facing web server:
the ops listener carries none of the Web machinery (TLS, CORS, WebSocket,
landing), gets an honest flag name, and parallels the existing `stats.rs`
config module. `render_metrics` + its test move with it; `/metrics` is gone
from the public web router, `/health` stays public there.

main.rs runs the service in its select!; it's a no-op (pending) when
`--metrics-listen` is unset. Embedders that call `Web::serve` must now spawn
`Metrics::run` themselves, since it's no longer bundled into the web server.
