+++ rs/moq-relay/src/lib.rs
@@ -12,6 +12,7 @@ mod cluster;
 mod config;
 mod connection;
 mod http_client;
+mod metrics;
 mod stats;
 mod web;
 #[cfg(feature = "websocket")]
@@ -25,5 +26,6 @@ pub use auth::*;
 pub use cluster::*;
 pub use config::*;
 pub use connection::*;
+pub use metrics::*;
 pub use stats::*;
 pub use web::*;
